Refactor filtering system, esp. by site type
FilterManager was moved to a separate module,
Filter was split into two classes, documented
and cleaned up:
- indentifier-value mapping was introduced for
  `choices` argument to enable use of non-
  serializable values
- advanced sqlalchemy support was improved,
  extended while the custom function interface
  was simplified

Site types and disease name filters we redesigned
taking advantage of the newly introduced Filter
capabilities.

A bug in mutations counting in gene browsing
tables was fixed, along with overhaul of the
query generation.

Index was created on sites-site_type association;
creation of indexes on other associations will be
easier now as `make_association_table()` accepts
boolean `index` argument now.

Minor improvements were made to "some values were
left out" warning flash.
